requirements.txt
djangorestframework djangorestframework-simplejwt
Django==3.2.12
django-crispy-forms==1.12.0
Pillow==9.0.0
django-cors-headers
